fix: move W3cExternalTraceAttributes to models
