+++ src/models/W3cExternalTraceAttributes.ts
@@ -1,0 +1,22 @@
+export type W3cExternalTraceAttributes = {
+  /**
+   * A key that determines if the traceparent header was found
+   */
+  isW3cHeaderFound: boolean | null;
+  /**
+   * A unique identifier for the trace generated by the SDK in case of no cought header found
+   */
+  partialId: number | null;
+  /**
+   * The start time of the network request
+   */
+  networkStartTimeInSeconds: number | null;
+  /**
+   * The traceparent header generated by the SDK
+   */
+  w3cGeneratedHeader: string | null;
+  /**
+   * The traceparent header received by the server
+   */
+  w3cCaughtHeader: string | null;
+};
